+++ app/Services/DTE/DTEFactory.php
@@ -3,7 +3,7 @@
 namespace App\Services\DTE;
 
 use App\Models\Invoice;
-//use App\Services\DTE\{ElectronicInvoice};
+use App\Services\DTE\Types\{ElectronicInvoice, ElectronicTicket};
 
 class DTEFactory
 {
@@ -12,6 +12,9 @@ class DTEFactory
         switch($type) {
             case 33:
                 return new ElectronicInvoice($invoice);
+                break;
+            case 39:
+                return new ElectronicTicket($invoice);
                 break;
         }
     }
