+++ app/Services/DTE/Traits/DTEArray.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Services\DTE\Traits;
+
+Trait DTEArray
+{
+
+}
