+++ resources/views/invoice/manage_invoice/go_to_edit.blade.php
@@ -1,3 +1,4 @@
-<a href="{{ url('admin/invoice/'.$invoice->id.'/edit') }} " class="btn btn-sm btn-primary">
+<a href="{{ url('admin/invoice/'.$invoice->id.'/edit') }}" class=" btn btn-sm btn-primary" > 
     <i class="la la-send"></i> Editar documento
 </a>
+
